+++ docs/CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
 - set potential-rejection label as additional MANUAL workflow condition
+
+### Fixed
+- fix the resolved affects check (OSIDB-5315)
 
 ## [5.13.1] - 2026-08-03
 ### Added